requirements.txt
# Core
fastapi>=0.128.0
uvicorn>=0.40.0
websockets>=15.0.1
elevenlabs==2.30.*
python-dotenv>=1.2.1
mutagen>=1.47.0

# Unit tests (not needed on prod)
pytest>=9.0.2
pytest-asyncio>=1.3.0
